Fix #1185: [sim plots] Refactor: rename to arima_dash.py, reduce DRY in sim view_elements
